// React/app/GerAdv_TS/ModelosDocumentos/Interfaces/interface.ModelosDocumentos.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
// copyright © 2000-2025 Menphis - Sistemas Inteligentes
// This file is part of the Source Genesys project                     
'use client';

export interface IModelosDocumentos {
// 202501251
    id: number;
 
	tipomodelodocumento: number,
	nome: string,
	remuneracao: string,
	assinatura: string,
	header: string,
	footer: string,
	extra1: string,
	extra2: string,
	extra3: string,
	outorgante: string,
	outorgados: string,
	poderes: string,
	objeto: string,
	titulo: string,
	testemunhas: string,
	css: string,
}

export interface IModelosDocumentosFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IModelosDocumentosIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}